feature methods namespaced within module
